bot/requirements.txt
aiogram==3.1.1
python-dotenv==1.0.0
pytz
# БД
fastapi-users[sqlalchemy]
SQLAlchemy==2.0.22
asyncpg==0.28.0
alembic==1.12.0
psycopg2-binary==2.9.9
redis==5.0.1